+++ src/test/java/fr/utc/miage/shares/ActionTest.java
@@ -93,7 +93,7 @@ class ActionTest {
         }
 
         @Override
-        public float valeur(final Jour aJour) {
+        public double valeur(final Jour aJour) {
             return 0.0F;
         }
     }
